+++ internal/cmd/company/iam/user/edit_test.go
@@ -28,46 +28,46 @@ import (
 	"github.com/stretchr/testify/require"
 )
 
-func TestAddUser(t *testing.T) {
+func TestEditUser(t *testing.T) {
 	testCases := map[string]struct {
 		server    *httptest.Server
 		companyID string
 		role      resources.ServiceAccountRole
-		userEmail string
+		userID    string
 		expectErr bool
 	}{
 		"create user": {
-			server:    addUserTestServer(t),
+			server:    editUserTestServer(t),
 			companyID: "success",
 			role:      resources.ServiceAccountRoleGuest,
-			userEmail: "user@example.com",
+			userID:    "000000000000000000000001",
 		},
 		"missing company": {
-			server:    addUserTestServer(t),
+			server:    editUserTestServer(t),
 			companyID: "",
 			role:      resources.ServiceAccountRoleGuest,
-			userEmail: "user@example.com",
+			userID:    "000000000000000000000001",
 			expectErr: true,
 		},
-		"missing email": {
-			server:    addUserTestServer(t),
+		"missing user id": {
+			server:    editUserTestServer(t),
 			companyID: "success",
 			role:      resources.ServiceAccountRoleGuest,
-			userEmail: "",
+			userID:    "",
 			expectErr: true,
 		},
 		"wrong role": {
-			server:    addUserTestServer(t),
+			server:    editUserTestServer(t),
 			companyID: "",
 			role:      resources.ServiceAccountRole("example"),
-			userEmail: "user@example.com",
+			userID:    "000000000000000000000001",
 			expectErr: true,
 		},
 		"error from backend": {
-			server:    addUserTestServer(t),
+			server:    editUserTestServer(t),
 			companyID: "fail",
 			role:      resources.ServiceAccountRoleCompanyOwner,
-			userEmail: "user@example.com",
+			userID:    "000000000000000000000001",
 			expectErr: true,
 		},
 	}
@@ -80,11 +80,11 @@ func TestAddUser(t *testing.T) {
 				Host: server.URL,
 			})
 			require.NoError(t, err)
-			err = addUserToCompany(
+			err = editCompanyUser(
 				context.TODO(),
 				client,
 				testCase.companyID,
-				testCase.userEmail,
+				testCase.userID,
 				testCase.role,
 			)
 
@@ -98,14 +98,14 @@ func TestAddUser(t *testing.T) {
 	}
 }
 
-func addUserTestServer(t *testing.T) *httptest.Server {
+func editUserTestServer(t *testing.T) *httptest.Server {
 	t.Helper()
 
 	return httptest.NewServer(http.HandlerFunc(func(w http.ResponseWriter, r *http.Request) {
 		switch {
-		case r.Method == http.MethodPost && r.URL.Path == fmt.Sprintf(addUserToCompanyTemplate, "success"):
+		case r.Method == http.MethodPatch && r.URL.Path == fmt.Sprintf(editUserRoleTemplate, "success", "000000000000000000000001"):
 			w.WriteHeader(http.StatusOK)
-		case r.Method == http.MethodPost && r.URL.Path == fmt.Sprintf(addUserToCompanyTemplate, "fail"):
+		case r.Method == http.MethodPatch && r.URL.Path == fmt.Sprintf(editUserRoleTemplate, "fail", "000000000000000000000001"):
 			w.WriteHeader(http.StatusBadRequest)
 		default:
 			require.Fail(t, "request not implemented", "request received for %s with %s method", r.URL, r.Method)