+++ CHANGELOG.md
@@ -14,6 +14,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Added
 
 - `company iam add user` command
+- `company iam edit user` command
 - `company iam add group` command
 - `company iam add group-member` command